Refactors MenuDropdown out of Card
