+++ app/(gcforms)/[locale]/(form administration)/forms/[[...path]]/components/client/MenuDropdownButton.tsx
@@ -2,19 +2,97 @@
 import {
   MenuDropdown,
   MenuDropdownItemI,
+  MenuDropdownItemCallback,
 } from "@clientComponents/myforms/MenuDropdown/MenuDropdown";
 import { useTranslation } from "react-i18next";
+import copy from "copy-to-clipboard";
+import { getForm } from "../../actions";
+import { getDate, slugify } from "@lib/client/clientHelpers";
+import { CardProps } from "../server/Card";
 
 export const MenuDropdownButton = ({
   id,
-  menuItemsList,
+  card,
   direction,
+  handleDelete,
 }: {
   id: string;
-  menuItemsList: MenuDropdownItemI[];
+  card: CardProps;
   direction: string;
+  handleDelete: (card: CardProps) => void;
 }) => {
-  const { t } = useTranslation(["my-forms", "common"]);
+  const { t, i18n } = useTranslation(["my-forms", "common"]);
+  const menuItemsList: Array<MenuDropdownItemI> = [
+    {
+      title: t("card.menu.save"),
+      callback: () => {
+        downloadForm(name, id);
+        return { message: "" };
+      },
+    },
+    {
+      title: t("card.menu.settings"),
+      url: `/${i18n.language}/form-builder/settings/${id}`,
+    },
+    {
+      title: t("card.menu.delete"),
+      callback: () => {
+        handleDelete(card);
+        return {
+          message: "",
+        };
+      },
+    },
+  ];
+
+  // Show slightly different list items depending on whether a Published or Draft card
+  if (card.isPublished) {
+    menuItemsList.unshift({
+      title: t("card.menu.copyLink"),
+      callback: copyLinkCallback,
+    });
+  } else {
+    menuItemsList.unshift(
+      {
+        title: t("card.menu.edit"),
+        url: `/${i18n.language}/form-builder/edit/${id}`,
+      },
+      {
+        title: t("card.menu.preview"),
+        url: `/${i18n.language}/form-builder/preview/${id} `,
+      }
+    );
+  }
+
+  async function downloadForm(name: string, id: string) {
+    const response = await getForm(id);
+    const fileName = name
+      ? name
+      : i18n.language === "fr"
+      ? response.form.titleFr
+      : response.form.titleEn;
+    const data = JSON.stringify(response.form, null, 2);
+    const tempUrl = window.URL.createObjectURL(new Blob([data]));
+    const link = document.createElement("a");
+    link.href = tempUrl;
+    link.setAttribute("download", slugify(`${fileName}-${getDate()}`) + ".json");
+    document.body.appendChild(link);
+    link.click();
+  }
+
+  function copyLinkCallback(): MenuDropdownItemCallback {
+    const path = `${window.location.origin}/${i18n.language}/id/${id}`;
+    if (copy(path)) {
+      return {
+        message: t("card.menu.coppiedToClipboard"),
+      };
+    }
+    return {
+      message: t("card.menu.somethingWentWrong"),
+      isError: true,
+    };
+  }
+
   return (
     <MenuDropdown id={id} items={menuItemsList} direction={direction}>
       <span className="mr-1 text-[2rem]" aria-hidden="true">
